Restructure Api calls
