+++ src/views/404View.vue
@@ -1,0 +1,9 @@
+<template>
+    error 404
+    <div class="btn" @click="keycloak.login()">Login</div>
+</template>
+
+<script setup lang="ts">
+import keycloak from '@/keycloak/keycloak';
+
+</script>
